Added fix to handle computer name changes after instance has been bound to Active Directory - EUCA-5921
The fix was to add in a check to see if the instance was a new launch.
If the instance is a new launch, then change the hostname.  If not,
don't change it.  This is important when users want to change the
computer name after the instance has been bound to Active Directory.
Thanks to Deependra Shekhawat for his help in solving this issue.