Fix for EUCA-5921 - support Windows computer name change
This fix is to handle the use cases for users who want to change the
computer name of an instance after it has been added to Active
Directory, or changing the computer name in general.  After changing the
computer name,  the Windows instance needs to be rebooted.  This fix
makes sure not to change the computer name if the instance is not newly
launched.  Thanks to Deependra Shekhawat for helping out with this
patch.